+++ src/index.html
@@ -131,7 +131,9 @@
 
 <script src="js/resizer.js" charset="utf-8"></script>
 <script src="js/upload.js" charset="utf-8"></script>
-<script src="js/pictures.js" charset="utf-8"></script>
+<!--<script src="js/load.js" charset="utf-8"></script>-->
+<!--<script src="js/getPictureTemplate.js" charset="utf-8"></script>-->
+<script src="js/main.js" charset="utf-8"></script>
 
 </body>
 </html>
